+++ sdk/identity/Azure.Identity/tests/StaticCachesUtilities.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using Microsoft.Identity.Client;
@@ -11,14 +10,28 @@ namespace Azure.Identity.Tests
 {
     internal static class StaticCachesUtilities
     {
-        private static readonly Lazy<Action> s_clearStaticMetadataProvider = new Lazy<Action>(() =>
+        private static readonly Action s_noOpAction = () => { };
+
+        internal static readonly Lazy<Action> s_clearStaticMetadataProvider = new Lazy<Action>(() =>
         {
-            Type staticMetadataProviderType = typeof(PublicClientApplication).Assembly.GetType("Microsoft.Identity.Client.Instance.Discovery.NetworkCacheMetadataProvider", true);
-            MethodInfo clearMethod = staticMetadataProviderType.GetMethod("Clear", BindingFlags.Public | BindingFlags.Instance);
-            NewExpression callConstructor = Expression.New(staticMetadataProviderType);
-            MethodCallExpression invokeClear = Expression.Call(callConstructor, clearMethod);
-            return Expression.Lambda<Action>(invokeClear).Compile();
+            Type staticMetadataProviderType = typeof(PublicClientApplication).Assembly.GetType("Microsoft.Identity.Client.Instance.Discovery.NetworkCacheMetadataProvider", false);
+            if (staticMetadataProviderType == null)
+            {
+                return s_noOpAction;
+            }
+
+            MethodInfo resetMethod = staticMetadataProviderType.GetMethod("ResetStaticCacheForTest", BindingFlags.NonPublic | BindingFlags.Static);
+            if (resetMethod == null)
+            {
+                return s_noOpAction;
+            }
+
+            return Expression.Lambda<Action>(Expression.Call(resetMethod)).Compile();
         });
-        public static void ClearStaticMetadataProviderCache() => s_clearStaticMetadataProvider.Value();
+
+        public static void ClearStaticMetadataProviderCache()
+        {
+            s_clearStaticMetadataProvider.Value();
+        }
     }
 }